ingestion/requirements.txt
azure-functions>=1.20.0
azure-identity>=1.18.0
azure-search-documents>=11.6.0
azure-storage-blob>=12.22.0
openai>=1.50.0
pypdf>=4.3.0
